teamd: use kernel port state when updating the published link state
Further to vrvdr-47776 & vrvdr-47598, there is still a small race
condition such that the published state of the port may not be up at
the point at which the master state is signalled as up.

Use the raw kernel interface state (ifi_flags) when determining if the
first link is up.

Ultimately the use of the controller PUB/SUB connection to collect
link state needs some re-design. The basic issue with each of these
"band-aids" is that the ZMQ link-watch thread connects to the
controller too late and misses the publication of the initial state
transition of the first link in the bundle. Are the updates from the
controller (still) necessary? Could the whole set of patches be
removed?

Fixes: VRVDR-50619